fix: change parameter type from uint32_t to size_t in mem_set function for consistency

+++ kernel/mem.c
@@ -25,7 +25,7 @@ void memory_copy(void *src, void *dest, size_t nbytes) {
     }
 }
 
-void mem_set(void *dst, uint8_t val, uint32_t count){
+void mem_set(void *dst, uint8_t val, size_t count){
     uint8_t *temp = (uint8_t *)dst;
     for( ; count != 0; count--) *temp++ = val;
 }
